Extract ListViewModel from Composables #1
